server/requirements.txt
bottle
paho-mqtt